backend/requirements.txt
Flask
flask-cors
google-cloud-storage
python-dotenv
pytesseract
Pillow
opencv-python-headless
sendgrid
gunicorn
certifi
